+++ back-end/src/app/controllers/order.controller.js
@@ -4,7 +4,7 @@ const { mongooseToObject } = require('../../utils/mongoose');
 
 class OrderController {
     getOrders(req, res) {
-        if (req.isAuthenticated()) {
+        if (req.isAuthenticated() && req.user.role === 'admin') {
             Order.find({})
                 .then(orders => {
                     res.json(multipleMongooseToObject(orders));
@@ -21,7 +21,7 @@ class OrderController {
     }
     
     getOrderById(req, res) {
-        if (req.isAuthenticated()) {
+        if (req.isAuthenticated() && req.user.role === 'admin') {
             Order.findById(req.params.id)
                 .populate({
                     path: 'products.product_id',
@@ -41,7 +41,7 @@ class OrderController {
     }
 
     createOrder(req, res) {
-        if (req.isAuthenticated()) {
+        if (req.isAuthenticated() && req.user.role === 'admin') {
             const order = new Order(req.body);
             order.save()
                 .then(order => {
@@ -57,7 +57,7 @@ class OrderController {
     }
 
     updateOrder(req, res) {
-        if (req.isAuthenticated()) {
+        if (req.isAuthenticated() && req.user.role === 'admin') {
             Order.findByIdAndUpdate(req.params.id, req.body, { new: true })
                 .then(order => {
                     res.json(mongooseToObject(order));
@@ -73,8 +73,25 @@ class OrderController {
         };
     }
 
-    deleteOrder(req, res) {
+    cancelOrder(req, res) {
         if (req.isAuthenticated()) {
+            Order.findByIdAndUpdate(req.params.id, { status: 'canceled' }, { new: true })
+                .then(order => {
+                    res.json(mongooseToObject(order));
+                }).catch(err => {
+                    res.json({
+                        message: err.message
+                    });
+                })
+        } else {
+            res.json({
+                message: 'You are not authorized'
+            });
+        };
+    }
+
+    deleteOrder(req, res) {
+        if (req.isAuthenticated() && req.user.role === 'admin') {
             Order.findByIdAndDelete(req.params.id)
                 .then(() => {
                     res.json({
